PRO-2950: move Context facade family under Axn::Core (Result stays public)
